move action tests
